moving common tests to a function
